requirements.txt
Flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.2.2
requests==2.33.1
nltk==3.9.4
numpy==1.26.4
scikit-learn==1.5.2
gunicorn==22.0.0
networkx==3.3
textblob==0.18.0